move video files by duration 3C.6.4
